+++ src/mqtt/time.rs
@@ -1,7 +1,6 @@
 use chrono::NaiveDateTime;
-use serde::Serialize;
 
-#[derive(Debug, Clone, Copy, Serialize)]
+#[derive(Debug, Clone, Copy, serde::Serialize)]
 #[serde(untagged)]
 pub enum Time {
     Retained,
@@ -26,7 +25,7 @@ impl Time {
     }
 
     #[cfg(test)]
-    pub fn datetime_example() -> NaiveDateTime {
+    pub const fn datetime_example() -> NaiveDateTime {
         chrono::NaiveDate::from_ymd_opt(1996, 12, 19)
             .unwrap()
             .and_hms_opt(16, 39, 57)